% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{FUNCTION_TITLE}
\usage{
age(birthDate, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthDate}{PARAM_DESCRIPTION}

\item{refDate}{PARAM_DESCRIPTION, Default: Sys.Date()}

\item{unit}{PARAM_DESCRIPTION, Default: 'year'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[lubridate:as.period]{as.period}, \link[lubridate:interval]{interval}
}
