% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwu_import_index_scores.R
\name{bwu_import_index_score}
\alias{bwu_import_index_score}
\title{Import/Tidy Excel Index Score File}
\usage{
bwu_import_index_score(patient)
}
\arguments{
\item{patient}{Name of patient}
}
\value{
A csv table
}
\description{
Import/Tidy Excel Index Score File
}
