% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwu_read_cognition_index_scores.R
\name{bwu_read_cognition_index_scores}
\alias{bwu_read_cognition_index_scores}
\title{Import Neurocognitive Index Scores}
\usage{
bwu_read_cognition_index_scores(patient)
}
\arguments{
\item{patient}{Name of patient}
}
\value{
A table
}
\description{
Import Neurocognitive Index Scores
}
