% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age}
\usage{
calc_age(birthDate, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthDate}{date of birth}

\item{refDate}{date of evaluation}

\item{unit}{unit year, month or day}
}
\value{
age
}
\description{
Calculate age at a given reference date
Create an interval between the date of birth and the enrollment date;
intervals are specific to the two dates. Periods give the actual length
of time between those dates, so convert to period and extract the year.
}
\examples{

dob <- "1977-07-13"
doe <- "2021-04-04"
calc_age(dob, doe)
}
