% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_ci_95}
\alias{calc_ci_95}
\title{Compute confidence intervals (CI)}
\usage{
calc_ci_95(x, m, sd, rel, round = 0, level = 0.95)
}
\arguments{
\item{x}{Score}

\item{m}{Mean of scale}

\item{sd}{Standard deviation (SD) of scale}

\item{rel}{Reliability of scale}

\item{round}{Number of digits to round to}

\item{level}{Confidence level}
}
\value{
Returns confidence interval
}
\description{
To compute confidence intervals for test scores.
}
\examples{
calc_ci_95(x = 88, m = 100, sd = 15, rel = .90, round = 0, level = 0.95)
}
