% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_percentile.R
\name{calc_percentile}
\alias{calc_percentile}
\title{Convert T scores to z scores to percentiles}
\usage{
calc_percentile(score, mean, sd)
}
\arguments{
\item{score}{Test score}

\item{mean}{Mean of whatever type of score}

\item{sd}{SD of whatever type of score}

\item{percentile}{percentile rank score}
}
\value{
percentile score
}
\description{
Convert T scores to z scores to percentiles
}
\examples{
calc_percentile(score = 103, mean = 100, sd = 15)
}
