% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_pred_score_ci_95_pct}
\alias{calc_pred_score_ci_95_pct}
\title{Predicted test score with confidence intervals}
\usage{
calc_pred_score_ci_95_pct(x, y, r, m, sd, rel, round = 0, level = 0.95)
}
\arguments{
\item{x}{Score 1 (ability)}

\item{y}{Score 2 (achievement)}

\item{r}{Correlation between ability test and achievement test}

\item{m}{Mean of test score}

\item{sd}{Standard deviation of test scores}

\item{rel}{Reliability of scale}

\item{round}{Number of digits to round to}

\item{level}{Confidence level}
}
\value{
A predicted score.
}
\description{
Estimate the difference between an actual score and a predicted score and compute the percentile of that score, and the confidence interval 95\% of the time.
}
\examples{
calc_predicted_ci(x = 122, y = 84, r = .65, m = 100, sd = 15, rel = .90, round = 0, level = 0.95)
}
