% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{cat_neuropsych_results}
\alias{cat_neuropsych_results}
\title{Concatenate and Flatten Neuropsych Results by Scale}
\usage{
cat_neuropsych_results(data, file, ...)
}
\arguments{
\item{data}{A dataframe containing the data}

\item{file}{A character string specifying the name of the file}

\item{...}{Additional arguments passed to other functions}
}
\value{
A file containing the flattened and scaled text
}
\description{
This function sorts the data by percentile, removes duplicates and converts the data to text. Finally, it appends the converted data to a file.
}
