% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_neuropsych_reports.R
\name{child_report}
\alias{child_report}
\title{Child report}
\usage{
child_report(
  fig_width = 4,
  fig_height = 2.5,
  fig_crop = TRUE,
  dev = "pdf",
  highlight = "default",
  ...
)
}
\description{
DESCRIPTION
}
