% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ci}
\alias{ci}
\title{Compute confidence intervals (CI)}
\usage{
ci(x, m, sd, rel, level = 0.95, ...)
}
\arguments{
\item{x}{Score}

\item{m}{Mean of scale}

\item{sd}{Standard deviation (SD) of scale}

\item{rel}{Reliability of scale}

\item{level}{Confidence level, default is 0.95}

\item{...}{Additional arguments}
}
\value{
Returns a named vector with elements 'true_score', 'ci_lo' and 'ci_hi'
}
\description{
To compute confidence intervals for test scores.
}
\examples{
ci_res <- ci(x = 88, m = 100, sd = 15, rel = .80)
}
