% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_95.R
\name{ci_95}
\alias{ci_95}
\title{Compute 95\% confidence intervals (CI)}
\usage{
ci_95(x, m, sd, rel, level = 0.95)
}
\arguments{
\item{x}{Score}

\item{m}{Mean of scale}

\item{sd}{Standard deviation (SD) of scale}

\item{rel}{Reliability of scale}

\item{level}{Confidence level}
}
\value{
Returns 95\% CI
}
\description{
To compute confidence intervals for test scores.
}
\examples{
ci_95(x = 88, m = 100, sd = 15, rel = .80)
}
