% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dots}
\alias{dots}
\title{Dot Counting Task from NIH EXAMINER
A dataset containing dot counting task data.
The variables are as follows:
raw age percentile n m md sd normValue}
\format{
A data frame with several rows and 8 variables:
\describe{
\item{raw}{raw score}
\item{age}{age in years}
\item{percentile}{percentile rank}
\item{n}{sample size per group}
\item{m}{mean score}
\item{md}{median score}
\item{sd}{standard deviation of score}
\item{normValue}{T score fot dot counting task}
}
}
\usage{
dots
}
\description{
Dot Counting Task from NIH EXAMINER
A dataset containing dot counting task data.
The variables are as follows:
raw age percentile n m md sd normValue
}
\keyword{datasets}
