% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{extract_and_save_transcript}
\alias{extract_and_save_transcript}
\title{Extract and save transcript}
\usage{
extract_and_save_transcript(x, file, begin, end, eol = "\\\\n")
}
\arguments{
\item{x}{A character vector of length 1 containing a file name or a URL.}

\item{file}{The path/name of the output file.}

\item{begin}{The pattern beginning the text to extract.}

\item{end}{The pattern ending the text to extract.}

\item{eol}{The line break character to add between paragraphs, Default: '\\n'.}
}
\value{
This function does not return any value.
}
\description{
This function extracts text between two patterns, adds line breaks to separate paragraphs (optional), and writes the formatted text to a markdown file.
}
\details{
This function reads the entire file as a single string, extracts the text between the patterns, adds line breaks to separate paragraphs (optional), and writes the formatted text to a markdown file.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
