% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_extract_text.R
\name{extract_text}
\alias{extract_text}
\title{Extract OtterAI Text}
\usage{
extract_text(patient, file_path, start_pattern, end_pattern)
}
\arguments{
\item{patient}{Name of patient}

\item{file_path}{Path to .txt/.md file}

\item{start_pattern}{First anchor}

\item{end_pattern}{Second anchor}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[stringr]{str_match}}, \code{\link[stringr]{str_trim}}
}
