% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{filter_domain}
\alias{filter_domain}
\title{Filters data by domain and scale}
\usage{
filter_domain(data, domain, filter_file)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{domain}{The domain name that the user wants to filter by}

\item{filter_file}{A text file containing a list of scales}
}
\value{
Returns a filtered data frame
}
\description{
Filters data by domain and scale
}
