% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_resource}
\alias{find_resource}
\title{FUNCTION_TITLE}
\usage{
find_resource(template, file)
}
\arguments{
\item{template}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
