% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{generate_g}
\alias{generate_g}
\title{Use Excel Index Scores to Create "g.csv"}
\usage{
generate_g(data, patient, scales, index_score_file)
}
\arguments{
\item{data}{Dataset}

\item{patient}{Patient name}

\item{scales}{A character vector of scale names}

\item{index_score_file}{Index score file}
}
\value{
A csv file of g
}
\description{
Get 'g' from index scores
}
