% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{generate_g_csv_from_index_scores}
\alias{generate_g_csv_from_index_scores}
\title{Make index scores to create g.csv.}
\usage{
generate_g_csv_from_index_scores(
  data,
  patient,
  scales = scales,
  index_score_file
)
}
\arguments{
\item{patient}{Patient name}

\item{scales}{A character vector of scale names}
}
\value{
A csv file of g
}
\description{
Get 'g' from index scores
}
