% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-plots.R
\name{ggdotplot}
\alias{ggdotplot}
\title{FUNCTION_TITLE}
\usage{
ggdotplot(
  data,
  x,
  y,
  combine = FALSE,
  merge = FALSE,
  color = "black",
  fill = "lightgray",
  palette = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  facet.by = NULL,
  panel.labs = NULL,
  short.panel.labs = TRUE,
  size = NULL,
  binwidth = NULL,
  select = NULL,
  remove = NULL,
  order = NULL,
  add = "mean_se",
  add.params = list(),
  error.plot = "pointrange",
  label = NULL,
  font.label = list(size = 11, color = "black"),
  label.select = NULL,
  repel = FALSE,
  label.rectangle = FALSE,
  ggtheme = theme_pubr(),
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{combine}{PARAM_DESCRIPTION, Default: FALSE}

\item{merge}{PARAM_DESCRIPTION, Default: FALSE}

\item{color}{PARAM_DESCRIPTION, Default: 'black'}

\item{fill}{PARAM_DESCRIPTION, Default: 'lightgray'}

\item{palette}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{xlab}{PARAM_DESCRIPTION, Default: NULL}

\item{ylab}{PARAM_DESCRIPTION, Default: NULL}

\item{facet.by}{PARAM_DESCRIPTION, Default: NULL}

\item{panel.labs}{PARAM_DESCRIPTION, Default: NULL}

\item{short.panel.labs}{PARAM_DESCRIPTION, Default: TRUE}

\item{size}{PARAM_DESCRIPTION, Default: NULL}

\item{binwidth}{PARAM_DESCRIPTION, Default: NULL}

\item{select}{PARAM_DESCRIPTION, Default: NULL}

\item{remove}{PARAM_DESCRIPTION, Default: NULL}

\item{order}{PARAM_DESCRIPTION, Default: NULL}

\item{add}{PARAM_DESCRIPTION, Default: 'mean_se'}

\item{add.params}{PARAM_DESCRIPTION, Default: list()}

\item{error.plot}{PARAM_DESCRIPTION, Default: 'pointrange'}

\item{label}{PARAM_DESCRIPTION, Default: NULL}

\item{font.label}{PARAM_DESCRIPTION, Default: list(size = 11, color = "black")}

\item{label.select}{PARAM_DESCRIPTION, Default: NULL}

\item{repel}{PARAM_DESCRIPTION, Default: FALSE}

\item{label.rectangle}{PARAM_DESCRIPTION, Default: FALSE}

\item{ggtheme}{PARAM_DESCRIPTION, Default: theme_pubr()}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
