% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-plots.R
\name{ggdotplot_core}
\alias{ggdotplot_core}
\title{FUNCTION_TITLE}
\usage{
ggdotplot_core(
  data,
  x,
  y,
  color = "black",
  fill = "lightgray",
  palette = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  size = NULL,
  dotsize = size,
  binwidth = NULL,
  add = "mean_se",
  add.params = list(),
  error.plot = "pointrange",
  position = position_dodge(0.8),
  ggtheme = theme_pubr(),
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{color}{PARAM_DESCRIPTION, Default: 'black'}

\item{fill}{PARAM_DESCRIPTION, Default: 'lightgray'}

\item{palette}{PARAM_DESCRIPTION, Default: NULL}

\item{title}{PARAM_DESCRIPTION, Default: NULL}

\item{xlab}{PARAM_DESCRIPTION, Default: NULL}

\item{ylab}{PARAM_DESCRIPTION, Default: NULL}

\item{size}{PARAM_DESCRIPTION, Default: NULL}

\item{dotsize}{PARAM_DESCRIPTION, Default: size}

\item{binwidth}{PARAM_DESCRIPTION, Default: NULL}

\item{add}{PARAM_DESCRIPTION, Default: 'mean_se'}

\item{add.params}{PARAM_DESCRIPTION, Default: list()}

\item{error.plot}{PARAM_DESCRIPTION, Default: 'pointrange'}

\item{position}{PARAM_DESCRIPTION, Default: position_dodge(0.8)}

\item{ggtheme}{PARAM_DESCRIPTION, Default: theme_pubr()}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
