% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_compute_percentile_range}
\alias{gpluck_compute_percentile_range}
\title{gpluck_compute_percentile_range}
\usage{
gpluck_compute_percentile_range(
  .x,
  .score = NA,
  .score_type = c("z_score", "scaled_score", "t_score", "standard_score"),
  percentile = NA,
  range = NA,
  pct1 = NA,
  pct2 = NA,
  pct3 = NA,
  z = NA,
  ...
)
}
\arguments{
\item{.x}{A numeric vector.}

\item{.score}{The value of scores to be calculated according to the scoring type. Default: NA}

\item{.score_type}{Type of scoring- it can be z_score, scaled_score, t_score or standard_score. Default: c("z_score", "scaled_score", "t_score", "standard_score")}

\item{percentile}{Percentile of x. Default: NA}

\item{range}{Difference between the highest and lowest percentiles. Default: NA}

\item{pct1}{First percentile. Default: NA}

\item{pct2}{Second percentile. Default: NA}

\item{pct3}{Third percentile. Default: NA}

\item{z}{Z-scores. Default: NA}

\item{...}{Other arguments not in use.}
}
\value{
Returns the score with the specified type.
}
\description{
Computes percentile, range, scores from vector .x and returns the score with the specified type.
}
\details{
Values are calculated from vector x according to the specified scoring type.
}
\examples{
\dontrun{
if (interactive()) {
  x <- c(10, 20, 30, 40, 50)
  gpluck_compute_percentile_range(x, .score_type = "z_score")
}
}
}
