% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_compute_percentile_range_v2}
\alias{gpluck_compute_percentile_range_v2}
\title{Compute Percentile and Range if Unknown Version 2}
\usage{
gpluck_compute_percentile_range_v2(
  .x,
  .score = NA,
  .score_type = c("z_score", "scaled_score", "t_score", "standard_score"),
  percentile = NA,
  range = NA,
  pct1 = NA,
  pct2 = NA,
  pct3 = NA,
  z = NA,
  ...
)
}
\arguments{
\item{.x}{A numeric vector.}

\item{.score}{The value of scores to be calculated according to the scoring type. Default: NA}

\item{.score_type}{Type of scoring- it can be z_score, scaled_score, t_score or standard_score. Default: c("z_score", "scaled_score", "t_score", "standard_score")}

\item{percentile}{Percentile of x. Default: NA}

\item{range}{Difference between the highest and lowest percentiles. Default: NA}

\item{pct1}{First percentile. Default: NA}

\item{pct2}{Second percentile. Default: NA}

\item{pct3}{Third percentile. Default: NA}

\item{z}{Z-scores. Default: NA}

\item{...}{Other arguments not in use.}
}
\value{
Returns the score with the specified type.
}
\description{
This function computes percentile, performance range from a provided numeric vector \code{.x} and returns the value according to the specified scoring type.
}
