% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R, R/gpluck_locate_areas.R
\name{gpluck_locate_areas}
\alias{gpluck_locate_areas}
\alias{gpluck_locate_areas2}
\title{Extract areas for PDF plucking of tables}
\usage{
gpluck_locate_areas(file, pages = NULL, ...)

gpluck_locate_areas2(file, pages = NULL, ...)
}
\arguments{
\item{file}{A character string specifying the path to a PDF file. This can also be a URL, in which case the file will be downloaded to the R temporary directory using \code{download.file}.}

\item{pages}{An optional integer vector specifying pages to extract from. To extract multiple tables from a given page, repeat the page number (e.g., \code{c(1,2,2,3)}).}

\item{\dots}{Other arguments passed to \code{\link{extract_tables}}.}
}
\description{
Interactively identify areas and extract

Interactively identify areas and extract
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

# locate areas only, using Shiny app
locate_areas(f)

# locate areas only, using native graphics device
locate_areas(f, widget = "shiny")

# locate areas and extract
extract_areas(f)
}
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

# locate areas only, using Shiny app
locate_areas(f)

# locate areas only, using native graphics device
locate_areas(f, widget = "shiny")

# locate areas and extract
extract_areas(f)
}
}
