% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R
\name{gpluck_extract_table}
\alias{gpluck_extract_table}
\title{FUNCTION_TITLE}
\usage{
gpluck_extract_table(
  file,
  pages = NULL,
  area = NULL,
  guess = FALSE,
  method = c("decide", "lattice", "stream"),
  output = c("matrix", "data.frame", "character", "asis", "csv", "tsv", "json"),
  ...
)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{pages}{PARAM_DESCRIPTION, Default: NULL}

\item{area}{PARAM_DESCRIPTION, Default: NULL}

\item{guess}{PARAM_DESCRIPTION, Default: FALSE}

\item{method}{PARAM_DESCRIPTION, Default: c("decide", "lattice", "stream")}

\item{output}{PARAM_DESCRIPTION, Default: c("matrix", "data.frame", "character", "asis", "csv", "tsv",
"json")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[tabulizer]{extract_tables}}
}
