% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_extract_tables}
\alias{gpluck_extract_tables}
\title{gpluck_extract_tablevscode-file://vscode-app/Applications/Visual\%20Studio\%20Code.app/Contents/Resources/app/out/vs/code/electron-sandbox/workbench/workbench.html}
\usage{
gpluck_extract_tables(
  file,
  pages = NULL,
  area = NULL,
  guess = FALSE,
  method = c("decide", "lattice", "stream"),
  output = c("matrix", "data.frame", "character", "asis", "csv", "tsv", "json"),
  ...
)
}
\arguments{
\item{file}{The path to the PDF file.}

\item{pages}{A single page number or vector of page numbers, Default: NULL}

\item{area}{An area on the page given as c(x0, y0, x1, y1). Default: NULL}

\item{guess}{Whether to attempt to detect tables when the coordinates are not given. Default: FALSE}

\item{method}{Method to use for parsing. Default: c("decide", "lattice", "stream")}

\item{output}{Output format (see \code{\link[tabulizer]{extract_tables}} for more details). Default: c("matrix", "data.frame", "character", "asis", "csv", "tsv", "json")}

\item{...}{Other arguments to \code{\link[tabulizer]{extract_tables}}}
}
\value{
A parsed table in either matrix, data frame, character, asis, csv, tsv or json format.
}
\description{
This function returns the tables in a PDF file as parsed by the tabulizer package.
}
\details{
This is a wrapper around the \code{\link[tabulizer]{extract_tables}} function that allows for easier access to the tables in a PDF document.
}
\seealso{
\code{\link[tabulizer]{extract_tables}}
}
