% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck_get_index_scores.R
\name{gpluck_get_index_score2}
\alias{gpluck_get_index_score2}
\title{Import Neurocognitive Index Scores}
\usage{
gpluck_get_index_score2(patient)
}
\arguments{
\item{patient}{Name of patient}
}
\value{
A table
}
\description{
Import Neurocognitive Index Scores
}
