% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R
\name{gpluck_get_index_scores}
\alias{gpluck_get_index_scores}
\title{Index scors.}
\usage{
gpluck_get_index_scores(patient)
}
\arguments{
\item{patient}{Patient name}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[readxl]{read_excel}}
\code{\link[here]{here}}
\code{\link[janitor]{clean_names}}
\code{\link[glue]{glue}}
\code{\link[readr]{write_delim}}
}
