% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_locate_areas}
\alias{gpluck_locate_areas}
\title{FUNCTION_TITLE}
\usage{
gpluck_locate_areas(file, pages = NULL, ...)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{pages}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[tabulizer]{locate_areas}}
}
