% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R
\name{gpluck_make_columns}
\alias{gpluck_make_columns}
\title{Make columns.}
\usage{
gpluck_make_columns(
  table,
  test,
  test_name,
  scale = NULL,
  raw_score = NULL,
  score = NULL,
  range = NULL,
  percentile = NULL,
  ci_95 = NULL,
  domain = c("Intelligence/General Ability", "Academic Skills", "Verbal/Language",
    "Visual Perception/Construction", "Attention/Executive", "Memory", "Motor",
    "Social Cognition", "Behavioral/Emotional/Social", "Personality Disorders",
    "Psychiatric Disorders", "Substance Use Disorders", "Psychosocial Problems", "ADHD",
    "Executive Dysfunction", "Effort/Validity", ""),
  subdomain = NULL,
  narrow = NULL,
  pass = c("Planning", "Attention", "Sequential", "Simultaneous", "Knowledge", NA),
  verbal = c("Verbal", "Nonverbal", NA),
  timed = c("Timed", "Untimed", NA),
  test_type = c("npsych_test", "rating_scale", "validity_indicator", "item", NA),
  score_type = c("raw_score", "scaled_score", "t_score", "standard_score", "z_score",
    "percentile", "base_rate", "beta_coefficient", NA),
  absort = NULL,
  description = NULL,
  result = NULL,
  ...
)
}
\arguments{
\item{table}{PARAM_DESCRIPTION}

\item{test}{PARAM_DESCRIPTION}

\item{test_name}{PARAM_DESCRIPTION}

\item{scale}{PARAM_DESCRIPTION, Default: NULL}

\item{raw_score}{PARAM_DESCRIPTION, Default: NULL}

\item{score}{PARAM_DESCRIPTION, Default: NULL}

\item{range}{PARAM_DESCRIPTION, Default: NULL}

\item{percentile}{PARAM_DESCRIPTION, Default: NULL}

\item{ci_95}{PARAM_DESCRIPTION, Default: NULL}

\item{domain}{PARAM_DESCRIPTION, Default: c("Intelligence/General Ability", "Academic Skills", "Verbal/Language",
"Visual Perception/Construction", "Attention/Executive",
"Memory", "Motor", "Social Cognition", "Behavioral/Emotional/Social",
"Personality Disorders", "Psychiatric Disorders", "Substance Use Disorders",
"Psychosocial Problems", "ADHD", "Executive Dysfunction",
"Effort/Validity", "")}

\item{subdomain}{PARAM_DESCRIPTION, Default: NULL}

\item{narrow}{PARAM_DESCRIPTION, Default: NULL}

\item{pass}{PARAM_DESCRIPTION, Default: c("Planning", "Attention", "Sequential", "Simultaneous", "Knowledge",
NA)}

\item{verbal}{PARAM_DESCRIPTION, Default: c("Verbal", "Nonverbal", NA)}

\item{timed}{PARAM_DESCRIPTION, Default: c("Timed", "Untimed", NA)}

\item{test_type}{PARAM_DESCRIPTION, Default: c("npsych_test", "rating_scale", "validity_indicator", "item",
NA)}

\item{score_type}{PARAM_DESCRIPTION, Default: c("raw_score", "scaled_score", "t_score", "standard_score", "z_score",
"percentile", "base_rate", "beta_coefficient", NA)}

\item{absort}{PARAM_DESCRIPTION, Default: NULL}

\item{description}{PARAM_DESCRIPTION, Default: NULL}

\item{result}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Make columns.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
