% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck_make_score_ranges.R
\name{gpluck_make_score_ranges}
\alias{gpluck_make_score_ranges}
\title{Make test score range (e.g., Below Average, Above Average)}
\usage{
gpluck_make_score_ranges(
  table = table,
  score = NULL,
  percentile = NULL,
  range = range,
  test_type,
  ...
)
}
\arguments{
\item{table}{name of table}

\item{score}{standard score}

\item{percentile}{percentile rank}

\item{range}{Range of performance score}

\item{test_type}{Type of test}

\item{...}{More if needed}
}
\value{
Returns a modified table.
}
\description{
Use a consistent set of ranges for performance
}
\examples{
neurocog <- gpluck_make_score_ranges(
table = neurocog,
 score = 50,
  percentile = 50,
   range = "",
    test_type = "npsych_test")

}
