% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck_t2z2p.R
\name{gpluck_t2z2p}
\alias{gpluck_t2z2p}
\title{Convert T-score to z-score to percentile}
\usage{
gpluck_t2z2p(..., table, score = NULL, score_type)
}
\arguments{
\item{...}{More expressions}

\item{table}{Name of data table}

\item{score}{T-score value}

\item{score_type}{Type of score}
}
\value{
A tidy csv file
}
\description{
Convert T-score to z-score to percentile
}
