% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsub_fixed}
\alias{gsub_fixed}
\title{FUNCTION_TITLE}
\usage{
gsub_fixed(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
