% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{hawaii}
\alias{hawaii}
\title{Hawaii Color Palette}
\format{
An object of class \code{character} of length 30.
}
\usage{
hawaii
}
\description{
A color palette named \code{hawaii}.
}
\examples{
barplot(1:10, col = hawaii[1:10])

}
\keyword{datasets}
