% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inherit_pdf_document}
\alias{inherit_pdf_document}
\title{FUNCTION_TITLE}
\usage{
inherit_pdf_document(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[bookdown]{html_document2}}
}
