% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_with_date.R
\name{knit_with_date}
\alias{knit_with_date}
\title{Custom Knit function for RStudio}
\usage{
knit_with_date(input, ...)
}
\arguments{
\item{input}{An input}

\item{...}{Additional arguments}
}
\description{
Custom Knit function for RStudio
}
