% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_neuropsych.R
\name{load_data}
\alias{load_data}
\title{This function reads .csv files of patient data and writes four different files of the same data that are categorized by neuropsychological test type.}
\usage{
load_data(patient)
}
\arguments{
\item{patient}{character, Name of patient, e.g. "Biggie"}
}
\value{
List with 4 elements, each element is a dataframe of patient data
}
\description{
This function reads .csv files of patient data and writes four different files of the same data that are categorized by neuropsychological test type.
}
