% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_dotplot_typ}
\alias{make_dotplot_typ}
\title{Make a Dotplot for Neurocognitive Domains in Typst
This function creates a dotplot figure for a given dataset.}
\usage{
make_dotplot_typ(
  data,
  pheno,
  x = data$z_mean_subdomain,
  y = data$subdomain,
  colors = NULL,
  return_plot = TRUE,
  filename = c("pdf", "png", "svg"),
  z_score_label,
  ...
)
}
\arguments{
\item{data}{Data.frame containing the input data}

\item{pheno}{Phenotype of interest}

\item{x}{Variable plotted on the x-axis, by default data$z_mean_subdomain}

\item{y}{Variable plotted on the y-axis, by default data$subdomain}

\item{colors}{A vector of colors to be used for the plot}

\item{return_plot}{Plot that is generated}

\item{filename}{Filename extension}

\item{z_score_label}{Label for the z-score}

\item{...}{Additional arguments passed to other functions}
}
\value{
A dotplot formatted for Typst
}
\description{
Make a Dotplot for Neurocognitive Domains in Typst
This function creates a dotplot figure for a given dataset.
}
