% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_fig}
\alias{make_fig}
\title{Make a dotplot for cognitive domain scores
This function creates a dotplot figure for a given dataset.}
\usage{
make_fig(data, x, y, pheno, colors = NULL)
}
\arguments{
\item{data}{Dataframe containing the input data}

\item{x}{Variable plotted on the x-axis, by default data_dotplot$z_mean_narrow}

\item{y}{Variable plotted on the y-axis, by default data_dotplot$narrow}

\item{pheno}{Phenotype of interest}

\item{colors}{A vector of colors to be used for the plot}
}
\value{
an object of class figure object
}
\description{
Make a dotplot for cognitive domain scores
This function creates a dotplot figure for a given dataset.
}
