% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_g_csv}
\alias{make_g_csv}
\title{make_g_csv}
\usage{
make_g_csv(patient)
}
\arguments{
\item{patient}{Root directory of patient}
}
\value{
Returns g index score
}
\description{
Gets index scores for given patient path
}
