% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{make_shades}
\alias{make_shades}
\title{Make shades
Given a color make \code{n} lighter or darker shades}
\usage{
make_shades(color, n, lighter = TRUE)
}
\arguments{
\item{color}{The color to make shades of}

\item{n}{The number of shades to make}

\item{lighter}{Whether to make lighter (\code{TRUE}) or darker (\code{FALSE})
shades}
}
\value{
A vector of \code{n} color hex codes
}
\description{
Make shades
Given a color make \code{n} lighter or darker shades
}
\examples{
# Five lighter shades
make_shades("goldenrod", 5)
# Five darker shades
make_shades("goldenrod", 5, lighter = FALSE)
}
