% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{margin_note}
\alias{margin_note}
\title{FUNCTION_TITLE}
\usage{
margin_note(text, icon = "&#8853;")
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{icon}{PARAM_DESCRIPTION, Default: '⊕'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
