% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_html.R
\name{margin_references}
\alias{margin_references}
\title{FUNCTION_TITLE}
\usage{
margin_references(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
