% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_html.R
\name{marginnote_html}
\alias{marginnote_html}
\title{FUNCTION_TITLE}
\usage{
marginnote_html(text = "", icon = "&#8853;")
}
\arguments{
\item{text}{PARAM_DESCRIPTION, Default: ''}

\item{icon}{PARAM_DESCRIPTION, Default: '⊕'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
