% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neurocog}
\alias{neurocog}
\title{neurocog data documentation}
\format{
A data frame with unknown rows and 26 or so variables:
\describe{
  \item{filename}{name of test file}
  \item{scale}{age in years}
  \item{score}{percentile rank}
  \item{percentile}{sample size per group}
  \item{range}{mean score}
  \item{ci_95}{median score}
  \item{raw_score}{standard deviation of score}
  \item{test}{T score fot dot counting task}
  \item{test_name}{T score fot dot counting task}
  \item{domain}{T score fot dot counting task}
  \item{subdomain}{T score fot dot counting task}
  \item{narrow}{T score fot dot counting task}
  \item{domain_broad}{T score fot dot counting task}
  \item{domain_test}{T score fot dot counting task}
  \item{test_type}{T score fot dot counting task}
  \item{score_type}{T score fot dot counting task}
  \item{timed}{T score fot dot counting task}
  \item{verbal}{T score fot dot counting task}
  \item{pass}{T score fot dot counting task}
  \item{absort}{T score fot dot counting task}
  \item{z}{T score fot dot counting task}
  \item{z_mean_dom}{T score fot dot counting task}
  \item{z_sd_dom}{T score fot dot counting task}
  \item{pct_mean_dom}{T score fot dot counting task}
  \item{pct_sd_dom}{T score fot dot counting task}
  \item{z_mean_sub}{T score fot dot counting task}
  \item{z_sd_sub}{T score fot dot counting task}
  \item{pct_mean_sub}{T score fot dot counting task}
  \item{pct_sd_sub}{T score fot dot counting task}
  \item{z_mean_narrow}{T score fot dot counting task}
  \item{z_sd_narrow}{T score fot dot counting task}
  \item{pct_mean_narrow}{T score fot dot counting task}
  \item{pct_sd_narrow}{T score fot dot counting task}
  \item{z_mean_verbal}{T score fot dot counting task}
  \item{z_sd_verbal}{T score fot dot counting task}
  \item{z_mean_pass}{T score fot dot counting task}
  \item{z_sd_pass}{T score fot dot counting task}
  \item{z_mean_timed}{T score fot dot counting task}
  \item{z_sd_timed}{T score fot dot counting task}
}
}
\usage{
neurocog
}
\description{
A dataset containing neuropsych data
 diamonds. The variables are as follows:
raw	age	percentile	n	m	md	sd	normValue
}
\keyword{datasets}
