% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_neuropsych_reports.R
\name{neuropsych_report}
\alias{neuropsych_report}
\title{Main report}
\usage{
neuropsych_report(
  documentclass = c("adult-report", "child-report"),
  fig_width = 4,
  fig_height = 2.5,
  fig_crop = TRUE,
  dev = "pdf",
  highlight = "default",
  template = template_resources("neuropsych_report", "neuro-report.tex"),
  ...
)
}
\description{
Add FUNCTION_DESCRIPTION
}
