% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropsych_report_adult.R
\name{neuropsych_report_adult}
\alias{neuropsych_report_adult}
\title{Rmarkdown Template for Neuropsychological Report Adult}
\usage{
neuropsych_report_adult(
  ...,
  number_sections = TRUE,
  md_extensions = c("-autolink_bare_uris"),
  csl = NULL,
  colorlinks = TRUE
)
}
\arguments{
\item{...}{arguments passed down to \code{\link[bookdown:html_document2]{bookdown::pdf_document2()}}}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{md_extensions}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{csl}{bibliography style in the \code{.csl} format}

\item{colorlinks}{should cross-references and links be
colored?}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{rmarkdown::render()}}
}
\description{
Rmarkdown Template for Neuropsychological Report Adult
}
\examples{

rmarkdown::draft("neuropsych-report.Rmd",
                 template = "neuropsych_report_adult",
                 package = "bwu",
                 edit = FALSE)
}
