% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pandoc2.0}
\alias{pandoc2.0}
\title{FUNCTION_TITLE}
\usage{
pandoc2.0()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[rmarkdown]{pandoc_available}}
}
