% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_html.R
\name{parse_footnotes}
\alias{parse_footnotes}
\title{FUNCTION_TITLE}
\usage{
parse_footnotes(x, fn_label = "fn")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{fn_label}{PARAM_DESCRIPTION, Default: 'fn'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
