% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{pct}
\alias{pct}
\title{FUNCTION_TITLE}
\usage{
pct(x, m, sd)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[stats:Normal]{pnorm}
}
