% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{pdf_02_filter_lines}
\alias{pdf_02_filter_lines}
\title{Filter lines in a PDF file}
\usage{
pdf_02_filter_lines(extracted_text, pdf_file, page = NULL, scale)
}
\arguments{
\item{extracted_text}{A \code{list} of strings, containing text extracted from the PDF file.}

\item{pdf_file}{Path to the PDF file.}

\item{page}{Optional \code{integer}: Page number in the PDF used for extraction, defaults to \code{NULL}.}

\item{scale}{The word which is searched in lines and then filtered.}
}
\value{
A \code{vector} of strings, for all lines that contains the word in \code{scale}.
}
\description{
Filter lines from text extracted from a PDF for given words.
}
