% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf_document_format}
\alias{pdf_document_format}
\title{FUNCTION_TITLE}
\usage{
pdf_document_format(..., format, template, csl = NULL, colorlinks = TRUE)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{format}{PARAM_DESCRIPTION}

\item{template}{PARAM_DESCRIPTION}

\item{csl}{PARAM_DESCRIPTION, Default: NULL}

\item{colorlinks}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[rmarkdown]{pandoc_path_arg}}
}
