% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pegboard_dom}
\alias{pegboard_dom}
\title{pegboard_dom dataset}
\format{
A list of 45 data frames, each with the following 7 variables:
\describe{
\item{raw}{Character vector of raw scores.}
\item{norm}{Numeric vector of norm scores.}
\item{percentile}{Numeric vector of percentile scores.}
\item{lowerCI}{Numeric vector of lower confidence interval scores.}
\item{upperCI}{Numeric vector of upper confidence interval scores.}
\item{lowerCI_PR}{Numeric vector of lower confidence interval PR scores.}
\item{upperCI_PR}{Numeric vector of upper confidence interval PR scores.}
}
}
\source{
\url{[INSERT SOURCE URL HERE]}
}
\usage{
pegboard_dom
}
\description{
A list of multiple data frames representing pegboard dominance data. Each data frame corresponds to a different age group.
There are 45 data frames in the list, each with 7 variables.
}
\references{
\url{[INSERT REFERENCES URL HERE]}
}
\keyword{datasets}
