% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{percentile}
\alias{percentile}
\title{Convert T scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks. That is about it.}
\usage{
percentile(score, mean, sd)
}
\arguments{
\item{score}{Raw test score}

\item{mean}{Mean of whatever type of score}

\item{sd}{Standard deviation (SD) of whatever type of score}
}
\value{
A percentile rank score
}
\description{
Convert T scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks. That is about it.
}
\examples{
percentile(score = 103, mean = 100, sd = 15)

}
