% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot colors}
\usage{
plot_colors(colors)
}
\arguments{
\item{colors}{Vector of colour to plot}
}
\value{
A ggplot2 object
}
\description{
Plot a vector of colours to see what they look like
}
\examples{
shades <- make_shades("goldenrod", 5)
plot_colors(shades)
}
