% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{pred}
\alias{pred}
\title{FUNCTION_TITLE}
\usage{
pred(x, y, r, m, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{r}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
