% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicted_score.R
\name{predicted_score}
\alias{predicted_score}
\title{Predicted Ability-Achievement Score}
\usage{
predicted_score(x, r, m = 100)
}
\arguments{
\item{x}{Actual score}

\item{r}{Correlation between ability test and achievement test}

\item{m}{Mean of test score}
}
\value{
A predicted score.
}
\description{
Estimate the difference between an ability and achievement score
}
\examples{
predicted_score(x = 82, r = .93, m = 100)

}
