% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quote_footer}
\alias{quote_footer}
\title{FUNCTION_TITLE}
\usage{
quote_footer(text)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
