% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{read_and_combine_files}
\alias{read_and_combine_files}
\title{Helper function to read and combine CSV files}
\usage{
read_and_combine_files(files)
}
\arguments{
\item{files}{Character vector of file paths}
}
\value{
Combined dataframe
}
\description{
Efficiently reads multiple CSV files and combines them into a single dataframe
}
\keyword{internal}
