% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npsych.R
\name{read_npsych}
\alias{read_npsych}
\title{Import neuropsych csv files/tests that do not have PDFs}
\usage{
read_npsych(patient, test)
}
\arguments{
\item{patient}{Name of patient}

\item{test}{Name of test}
}
\value{
csv with new column 'result'
}
\description{
Import neuropsych csv files/tests that do not have PDFs
}
