% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{read_write_transcript_otterai}
\alias{read_write_transcript_otterai}
\title{Extract and Save NSE Transcript from Otter.ai}
\usage{
read_write_transcript_otterai(input, output, begin, end, eol = "\\\\n")
}
\arguments{
\item{input}{A character string specifying the path to the input file.}

\item{output}{A character string specifying the path to the output file where the extracted transcript will be saved.}

\item{begin}{A character string specifying the pattern that marks the beginning of the transcript in the input file.}

\item{end}{A character string specifying the pattern that marks the end of the transcript in the input file.}

\item{eol}{A character string specifying the line ending to be used in the output file, default is '\\n'.}
}
\value{
NULL, as this function is primarily used for its side effect of writing a file.
}
\description{
This function extracts the text between given beginning and ending patterns from a file and saves the transcript to another file.
}
\details{
This function uses regular expressions to extract text between specified patterns. It is useful for extracting structured text data, such as transcripts, from files.
}
