% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render REport}
\usage{
render_report(
  patient = "Biggie",
  input,
  output_format = NULL,
  params = NULL,
  ...
)
}
\arguments{
\item{patient}{Patient name.}

\item{input}{Input.}

\item{output_format}{adult or child.}

\item{params}{Parameters or "ask".}

\item{...}{Patient name, etc.}
}
\value{
RMarkdown report
}
\description{
Not sure
}
\seealso{
\link[rmarkdown:render]{render}
\link[bookdown:html_document2]{pdf_document2}
}
