% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocft.R
\name{rocft_copy}
\alias{rocft_copy}
\title{FUNCTION_TITLE}
\usage{
rocft_copy(raw_score, age)
}
\arguments{
\item{raw_score}{PARAM_DESCRIPTION}

\item{age}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
