% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropsych_test_scoring.R
\name{rocft_copy_t_score}
\alias{rocft_copy_t_score}
\title{Rey Complex Figure, Copy Trial Score Lookup}
\usage{
rocft_copy_t_score(age, raw_score)
}
\arguments{
\item{age}{A numeric value of the age of the participant. Must be between 16 and 89 years old.}

\item{raw_score}{A numeric value of the raw score from the Rey Complex Figure, Copy trial.}
}
\value{
Returns a numeric value of the T-score for the Rey Complex Figure, Copy trial.
}
\description{
Calculates a Rey Complex Figure, Copy trial score (T-score) from a raw score and age. The predicted score and standard deviation are also calculated.
}
\details{
This function takes the raw score and age of a participant and calculates the Rey Complex Figure, Copy trial T-score using the formula provided by \emph{Manly et al. (2000)}. It also returns the predicted score and standard deviation for the trial.
}
\references{
Manly, J.J., Schinka, J.A., & Guerrero, L. (2000). Normative Data on the Rey Complex Figure Test in Older Adults. Archives of Clinical Neuropsychology, 15(7), 613-621.
}
