% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocft.R
\name{rocft_delay}
\alias{rocft_delay}
\title{ROCFT Delayed Recall Standardized Score}
\usage{
rocft_delay(raw_score, age)
}
\arguments{
\item{raw_score}{Patients raw score on Delayed Recall}

\item{age}{Patient's age}
}
\value{
A T-Score of ROCFT Delayed Recall performance
}
\description{
ROCFT predicted scores, for age range of 22 -79 years, for
long-term delayed recall.
}
\examples{
rocft_delay_score <- rocft_delay(raw_score = 14, age = 24)
}
