% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocft.R
\name{rocftCopy}
\alias{rocftCopy}
\title{ROCFT Copy Standardized Score}
\usage{
rocftCopy(rawScore, age)
}
\arguments{
\item{rawScore}{Patients raw score}

\item{age}{Patients age}
}
\value{
A printed T-Score of ROCFT Copy performance
}
\description{
ROCFT Copy Standardized Score
}
\examples{
rocft_copy <- rocftCopy(rawScore = 32, age = 24)
}
