% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocft.R
\name{rocftDelay}
\alias{rocftDelay}
\title{ROCFT Delayed Recall Standardized Score}
\usage{
rocftDelay(rawScore, age)
}
\arguments{
\item{rawScore}{Patients raw score on Delayed Recall}

\item{age}{Patient's age}
}
\value{
A T-Score of ROCFT Delayed Recall performance
}
\description{
ROCFT Delayed Recall Standardized Score
}
\examples{
rocft_delay <- rocftDelay(rawScore = 14, age = 24)
}
