% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{roma}
\alias{roma}
\title{Roma Color Palette}
\format{
An object of class \code{character} of length 50.
}
\usage{
roma
}
\description{
A color palette named \code{roma}.
}
\examples{
barplot(1:10, col = roma[1:10])

}
\keyword{datasets}
