% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{roma_palette}
\alias{roma_palette}
\title{Create a color palette from the Roma palette}
\usage{
roma_palette(n)
}
\arguments{
\item{n}{The number of colors to sample from the Roma palette.}
}
\value{
A vector of colors sampled from the Roma palette.
}
\description{
Create a color palette from the Roma palette containing 24 colors. If n is greater than the length of roma, return the full palette.
}
\examples{
\dontrun{
roma_palette(3)
}

}
