% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sans_serif}
\alias{sans_serif}
\title{FUNCTION_TITLE}
\usage{
sans_serif(text)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
