% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_extract_text.R
\name{save_as_markdown}
\alias{save_as_markdown}
\title{Save extracted text as markdown.}
\usage{
save_as_markdown(text, output_path)
}
\arguments{
\item{text}{Name of file.}

\item{output_path}{Where to save outoput.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
