% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scico_palette}
\alias{scico_palette}
\title{Create a color palette from the Scico palette}
\usage{
scico_palette(n, palette = c("roma", "hawaii", "tokyo", "vik"))
}
\arguments{
\item{n}{The number of colors to sample from the Roma palette.}
}
\value{
A vector of colors sampled from the Roma palette.
}
\description{
Create a color palette from the Roma palette containing 24 colors. If n is greater than the length of roma, return the full palette.
}
\examples{
\dontrun{
scico_palette(3)
}

}
