% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sinewOxy}
\alias{sinewOxy}
\title{FUNCTION_TITLE}
\usage{
sinewOxy()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[sinew]{makeOxyFile}}
\code{\link[here]{here}}
}
