% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{standardization}
\alias{standardization}
\title{Standardization of a vector}
\usage{
standardization(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
The standardization of the vector.
}
\description{
This function takes a vector as input and returns the standardization of that vector.
}
\examples{
x <- c(1, 2, 3)
standardization(x)

}
