% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2z2p.R
\name{t2z2p}
\alias{t2z2p}
\title{Convert T-score to z-score to percentile}
\usage{
t2z2p(..., table, score)
}
\arguments{
\item{...}{More expressions}

\item{table}{Name of data table}

\item{score}{T-score value}
}
\value{
A tidy csv file
}
\description{
Convert T-score to z-score to percentile
}
