% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_gt}
\alias{table_gt}
\title{Make Table Using \code{gt} Package for Neurocognitive Domains v2}
\usage{
table_gt(
  data,
  pheno = NULL,
  table_name = NULL,
  source_note = NULL,
  title = NULL,
  tab_stubhead = NULL,
  caption = NULL,
  process_md = FALSE,
  vertical_padding = 0.1,
  groups_standard_score = c("NAB"),
  groups_t_score = c("NAB", "NIH EXAMINER", "Trail Making Test"),
  groups_scaled_score = c("WAIS-IV"),
  groups_z_score = c("WAIS-IV"),
  multiline = TRUE,
  ...
)
}
\arguments{
\item{data}{File or path to data.}

\item{pheno}{Phenotype name.}

\item{table_name}{Name of the table to be saved.}

\item{source_note}{Source note to be added to the table.}

\item{title}{Title of the table.}

\item{tab_stubhead}{Stubhead of the table.}

\item{caption}{Caption of the table.}

\item{process_md}{Process markdown.}

\item{vertical_padding}{Vertical padding.}

\item{groups_standard_score}{Groups for standard score.}

\item{groups_t_score}{Groups for t score.}

\item{groups_scaled_score}{Groups for scaled score.}

\item{groups_z_score}{Groups for raw score.}

\item{multiline}{Multiline footnotes, Default = TRUE.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A formatted table with domain counts.
}
\description{
Create a table of domain counts using dplyr and gt packages.
}
\details{
This function creates a table of domain counts from a data frame
using the dplyr and gt packages. It also saves the table with the specified
name. The \strong{table_gt} function can be used to create a table of domain
counts from a data frame using the dplyr and gt packages.
The output is a formatted table with domain counts along with the specified table name as a PDF file.
The function also supports additional arguments which are passed to it.
}
