% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_kbl}
\alias{tbl_kbl}
\title{Create Kable Table from Data}
\usage{
tbl_kbl(data)
}
\arguments{
\item{data}{The data frame to be used for creating the kable table.}
}
\value{
A kable table in LaTeX format.
}
\description{
This function creates a kable table from a given data frame.
}
\details{
This function uses the kableExtra package to create the kable table with the following specifications: lightable_options = "basic", latex_options = c("scale_down", "HOLD_position", "striped"), width of the first column is set to 8cm, rows are packed based on a table of the data's Test column, and the first row is set to bold. Additionally, a footnote is added at the end of the table.
}
\examples{
\dontrun{
if (interactive()) {
  df_adhd_ef <- read.csv("data/df_adhd_ef.csv")
  tbl_kbl(df_adhd_ef)
}
}
}
\seealso{
\code{\link[kableExtra]{kbl}}, \code{\link[kableExtra]{kable_classic}}, \code{\link[kableExtra]{kable_styling}}, \code{\link[kableExtra]{column_spec}}, \code{\link[kableExtra]{group_rows}}, \code{\link[kableExtra]{row_spec}}, \code{\link[kableExtra]{footnote}}
}
