% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_resources}
\alias{template_resources}
\title{FUNCTION_TITLE}
\usage{
template_resources(name, ...)
}
\arguments{
\item{name}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
