% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropsych_test_scoring.R
\name{tmt_a}
\alias{tmt_a}
\title{TMT, Part A Score Lookup}
\usage{
tmt_a(raw_score, age)
}
\arguments{
\item{raw_score}{The raw score inputted}

\item{age}{The age of the subject}
}
\value{
A string containing the T-score, z-score, predicted score, and predicted standard deviation
}
\description{
Calculates the TMT predicted score for Trails A with ages between 16 and 89
}
