% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scoring.R
\name{tmt_b}
\alias{tmt_b}
\title{TMT, B scoring function}
\usage{
tmt_b(raw_score, age)
}
\arguments{
\item{raw_score}{Score on the Trail Making Test B}

\item{age}{Age of the participant}

\item{t_score}{T-Score for TMT, B}
}
\value{
T-Score for TMT, B
}
\description{
Trails B Ages 16 - 89
}
