% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{tokyo}
\alias{tokyo}
\title{Tokyo Color Palette}
\format{
An object of class \code{character} of length 30.
}
\usage{
tokyo
}
\description{
A color palette named \code{tokyo}.
}
\examples{
barplot(1:10, col = tokyo[1:10])

}
\keyword{datasets}
