% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_handout.R
\name{tufte_book}
\alias{tufte_book}
\title{FUNCTION_TITLE}
\usage{
tufte_book(
  fig_width = 4,
  fig_height = 2.5,
  fig_crop = TRUE,
  dev = "pdf",
  highlight = "default",
  ...
)
}
\arguments{
\item{fig_width}{PARAM_DESCRIPTION, Default: 4}

\item{fig_height}{PARAM_DESCRIPTION, Default: 2.5}

\item{fig_crop}{PARAM_DESCRIPTION, Default: TRUE}

\item{dev}{PARAM_DESCRIPTION, Default: 'pdf'}

\item{highlight}{PARAM_DESCRIPTION, Default: 'default'}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
