% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_html.R
\name{tufte_html}
\alias{tufte_html}
\title{FUNCTION_TITLE}
\usage{
tufte_html(
  ...,
  tufte_features = c("fonts", "background", "italics"),
  tufte_variant = c("default", "envisioned"),
  margin_references = TRUE
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{tufte_features}{PARAM_DESCRIPTION, Default: c("fonts", "background", "italics")}

\item{tufte_variant}{PARAM_DESCRIPTION, Default: c("default", "envisioned")}

\item{margin_references}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
