% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tufte_html.R
\name{tufte_html_dependency}
\alias{tufte_html_dependency}
\title{FUNCTION_TITLE}
\usage{
tufte_html_dependency(features, variant)
}
\arguments{
\item{features}{PARAM_DESCRIPTION}

\item{variant}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
