% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_drive.R
\name{upload_to_drive}
\alias{upload_to_drive}
\title{Upload a file to Google Drive}
\usage{
upload_to_drive(file_path, file_name)
}
\arguments{
\item{file_path}{The local path of the file to be uploaded.}

\item{file_name}{The name to assign to the file once uploaded on Google Drive.}
}
\value{
The function returns an object of class \code{drive_file} which contains
metadata of the uploaded file.
}
\description{
This function uploads a specified file to Google Drive under a specified file name.
It requires the googledrive package and proper authentication to access Google Drive.
}
\examples{
# Make sure to replace '/path/to/your/file.csv' and 'file_name.csv'
# with actual values. Uncomment and run:
# upload_to_drive("/path/to/your/file.csv", "file_name.csv")

}
\keyword{cloud}
\keyword{storage}
