% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{vik}
\alias{vik}
\title{Vik Color Palette}
\format{
An object of class \code{character} of length 21.
}
\usage{
vik
}
\description{
A color palette named \code{vik}.
}
\examples{
barplot(1:10, col = vik[1:10])

}
\keyword{datasets}
