% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{write_domain_csv}
\alias{write_domain_csv}
\title{Write Domain CSV}
\usage{
write_domain_csv(data, pheno)
}
\arguments{
\item{data}{Numeric vector, data frame, or matrix.}

\item{pheno}{Character vector of strings.}
}
\description{
Write Domain CSV
}
