% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_neuropsych_reports.R
\name{adult_report}
\alias{adult_report}
\title{BWU Neuropsych Report formats (child and adult)}
\usage{
adult_report(
  fig_width = 4,
  fig_height = 2.5,
  fig_crop = TRUE,
  dev = "pdf",
  highlight = "default",
  ...
)
}
\arguments{
\item{...}{Other arguments to be passed to \code{\link[=pdf_document2]{pdf_document2()}} or
\code{\link[=html_document]{html_document()}} (note you cannot use the \code{template}
argument in \code{tufte_handout} or the \code{theme} argument in
\code{tufte_html()}; these arguments have been set internally)}
}
\description{
Templates for creating handouts according to the style of Edward R. Tufte and
Richard Feynman.
}
\examples{
library(bwu)
}
\references{
See \url{https://rstudio.github.io/tufte/} for an example.
}
