% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Calculate age}
\usage{
age(birthDate, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthDate}{Date of birth}

\item{refDate}{Date of evaluation}

\item{unit}{Unit of time (year, month or day)}
}
\value{
Age in years.
}
\description{
Calculate age at a given reference date
Create an interval between the date of birth and the enrollment date;
intervals are specific to the two dates. Periods give the actual length
of time between those dates, so convert to period and extract the year.
}
\examples{
dob <- "1977-07-13"
doe <- "2021-04-04"
age(dob, doe)

}
