% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Compute age}
\usage{
age(birthDate, refDate = Sys.Date(), unit = "year")
}
\arguments{
\item{birthDate}{Date of birth}

\item{refDate}{Reference date, Default: Sys.Date()}

\item{unit}{Date unit, Default: 'year'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Compute age of someone.
}
\seealso{
\link[lubridate:as.period]{as.period}, \link[lubridate:interval]{interval}
}
