% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci_95.R
\name{calc_ci_95}
\alias{calc_ci_95}
\title{Compute 95\% CI}
\usage{
calc_ci_95(x, m, sd, rel, level = 0.95)
}
\arguments{
\item{x}{score}

\item{m}{mean of scale}

\item{sd}{sd of scale}

\item{rel}{reliability of scale}

\item{level}{confidence level}
}
\value{
Returns 95\% CI
}
\description{
To compute confidence interval.
}
\examples{
ci_95(x = 88, m = 100, sd = 15, rel = .80)

}
