% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_ci_95}
\alias{calc_ci_95}
\title{Compute 95\% Confidence Intervals}
\usage{
calc_ci_95(
  ability_score,
  mean,
  standard_deviation,
  reliability,
  round = 0,
  confidence_level = 0.95
)
}
\arguments{
\item{ability_score}{Score}

\item{mean}{Mean of scale}

\item{standard_deviation}{Standard deviation (SD) of scale}

\item{reliability}{Reliability of scale}

\item{round}{Number of digits to round to}

\item{confidence_level}{Confidence level}
}
\value{
Returns confidence interval
}
\description{
To compute confidence intervals for test scores.
}
