% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_percentile}
\alias{calc_percentile}
\title{Convert SS/T/scaled scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks.}
\usage{
calc_percentile(x, m, sd, round = 0)
}
\arguments{
\item{x}{Raw test score}

\item{m}{Mean of whatever type of score}

\item{sd}{Standard deviation (SD) of whatever type of score}

\item{round}{Number of digits to round to}
}
\value{
A percentile rank score
}
\description{
Convert SS/T/scaled scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks.
}
\examples{
calc_percentile(x = 103, m = 100, sd = 15, round = 1)
}
