% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_percentile}
\alias{calc_percentile}
\title{Convert Standard Score/T-Score/Scaled Score to z-Scores to Percentiles
This is a functioning that converts T-scores to z-scores to percentile
ranks.}
\usage{
calc_percentile(ability_score, mean, standard_deviation, round = 0)
}
\arguments{
\item{ability_score}{Raw test score}

\item{mean}{Mean of whatever type of score}

\item{standard_deviation}{Standard deviation (SD) of whatever type of score}

\item{round}{Number of digits to round to}
}
\value{
A percentile rank score
}
\description{
Convert Standard Score/T-Score/Scaled Score to z-Scores to Percentiles
This is a functioning that converts T-scores to z-scores to percentile
ranks.
}
\examples{
calc_percentile(ability_score = 103, mean = 100, standard_deviation = 15, round = 1)
}
