% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_predicted_score}
\alias{calc_predicted_score}
\title{Predicted test score}
\usage{
calc_predicted_score(x, y, r, m, round = 0)
}
\arguments{
\item{x}{Score 1 (ability)}

\item{y}{Score 2 (achievement)}

\item{r}{Correlation between ability test and achievement test}

\item{m}{Mean of test score}

\item{round}{Number of digits to round to}
}
\value{
A predicted score.
}
\description{
Estimate the difference between an actual score and a predicted score.
}
\examples{
calc_predicted_score(x = 122, y = 84, r = .65, m = 100)
}
