% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{calc_predicted_score}
\alias{calc_predicted_score}
\title{Predicted Test Score}
\usage{
calc_predicted_score(
  ability_score,
  achievement_score,
  correlation,
  mean,
  round = 0
)
}
\arguments{
\item{ability_score}{Score 1 (ability)}

\item{achievement_score}{Score 2 (achievement)}

\item{correlation}{Correlation between ability test and achievement test}

\item{mean}{Mean of test score}

\item{round}{Number of digits to round to}
}
\value{
A predicted score.
}
\description{
Estimate the difference between an actual score and a predicted score.
}
\examples{
calc_predicted_score(ability_score = 122, achievement_score = 84, correlation = .65, mean = 100)
}
