% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ci}
\alias{ci}
\title{FUNCTION_TITLE}
\usage{
ci(x, m, sd, rel, level = 0.95, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{m}{PARAM_DESCRIPTION}

\item{sd}{PARAM_DESCRIPTION}

\item{rel}{PARAM_DESCRIPTION}

\item{level}{PARAM_DESCRIPTION, Default: 0.95}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[stats:Normal]{qnorm}
}
