% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{compute_pctile_range}
\alias{compute_pctile_range}
\title{compute_pctile_range}
\usage{
compute_pctile_range(
  .x,
  .score = NA,
  .score_type = c("z_score", "scaled_score", "t_score", "standard_score", "raw_score"),
  percentile = NA,
  range = NA,
  pct1 = NA,
  pct2 = NA,
  pct3 = NA,
  z = NA,
  ...
)
}
\arguments{
\item{.x}{a numeric vector of scores}

\item{.score}{a single numeric score for which percentiles are to be computed, Default: NA}

\item{.score_type}{type of score, Default: c("z_score", "scaled_score", "t_score", "standard_score", "raw_score")}

\item{percentile}{numeric vector of percentile rank, Default: NA}

\item{range}{numeric vector of range, Default: NA}

\item{pct1}{1st percentile, Default: NA}

\item{pct2}{2nd percentile, Default: NA}

\item{pct3}{3rd percentile, Default: NA}

\item{z}{Z-score, Default: NA}

\item{...}{additional arguments passed to \code{pnorm} function}
}
\value{
returns list of percentile, range, and percentiles
}
\description{
Computes percentile, range, and percentiles for a given score value
}
\details{
Provides appropriate percentiles based on the score argument supplied. It covers following types
of scores such as z-score, t-score, scaled, raw and standard
}
\examples{
\dontrun{
if (interactive()) {
  scores <- c(35, 52, 74, 98)
  compute_pctile_range(scores, 74) # this will return percentile, range, pct1, pct2 and pct3
}
}
}
