% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dotplot}
\alias{dotplot}
\title{Create a dotplot of cognitive and behavioral domains.}
\usage{
dotplot(
  data,
  x,
  y,
  linewidth = 0.5,
  fill = x,
  shape = 21,
  point_size = 6,
  line_color = "black",
  colors = NULL,
  theme = "fivethirtyeight",
  return_plot = TRUE,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{The data frame containing the data for the dotplot.}

\item{x}{The column name in the data frame for the x-axis variable, typically
the mean z-score for a cognitive domain.}

\item{y}{The column name in the data frame for the y-axis variable, typically
the cognitive domain to plot.}

\item{linewidth}{The width of the line, Default: 0.5}

\item{fill}{The fill color for the points, Default: x-axis variable}

\item{shape}{The shape of the points, Default: 21}

\item{point_size}{The size of the points, Default: 6}

\item{line_color}{The color of the lines, Default: 'black'}

\item{colors}{A vector of colors for fill gradient, Default: NULL (uses pre-defined color palette)}

\item{theme}{The ggplot theme to be used, Default: 'fivethirtyeight'. Other options include 'minimal' and 'classic'}

\item{return_plot}{Whether to return the plot object, Default: TRUE}

\item{filename}{The filename to save the plot to, Default: NULL}

\item{...}{Additional arguments to be passed to the ggplot function.}
}
\value{
An object of class 'ggplot' representing the dotplot.
}
\description{
This function generates a dot plot with the given data, x and y. The point's aesthetics and plot theme can be customized.
}
\details{
This function generates a dot plot with the given data, x and y. The points' aesthetics and plot theme can be customized.
}
\examples{
# To get the plot object:
p <- dotplot(data, x, y)
# To save the plot as a PNG and not return the plot object:
dotplot(data, x, y, return_plot = FALSE, filename = "plot.png")
# To get the plot object and also save it:
p <- dotplot(data, x, y, filename = "plot.png")
}
