% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Make a dotplot for cognitive domains}
\usage{
dotplot(data, x, y, ...)
}
\arguments{
\item{data}{Name of dataset}

\item{x}{Name of x-axis variable}

\item{y}{Name of y-axis variable}

\item{...}{Additional arguments}
}
\value{
A dotplot using ggplot2
}
\description{
Description of the function.
}
\examples{
dotplot(iris, iris$Sepal.Length, iris$Species)

}
