% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{FUNCTION_TITLE}
\usage{
dotplot(
  data,
  x,
  y,
  linewidth = 0.5,
  fill = x,
  shape = 21,
  size = 6,
  color = "black",
  colors = NULL
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{linewidth}{PARAM_DESCRIPTION, Default: 0.5}

\item{fill}{PARAM_DESCRIPTION, Default: x}

\item{shape}{PARAM_DESCRIPTION, Default: 21}

\item{size}{PARAM_DESCRIPTION, Default: 6}

\item{color}{PARAM_DESCRIPTION, Default: 'black'}

\item{colors}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[ggplot2:ggplot]{ggplot}, \link[ggplot2:geom_segment]{geom_segment}, \link[ggplot2:aes]{aes}, \link[ggplot2:geom_point]{geom_point}, \link[ggplot2:scale_gradient]{scale_fill_gradientn}, \link[ggplot2:theme]{theme}, \link[ggplot2:element]{element_rect}
\link[stats:reorder.factor]{reorder}
\link[ggthemes:theme_fivethirtyeight]{theme_fivethirtyeight}
}
