% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Create a dotplot}
\usage{
dotplot(
  data,
  x,
  y,
  linewidth = 0.5,
  fill,
  shape = 21,
  point_size = 6,
  line_color = "black",
  colors = NULL,
  theme = "fivethirtyeight"
)
}
\arguments{
\item{data}{The data frame containing the data for the dot plot.}

\item{x}{The column name in the data frame for the x-axis variable.}

\item{y}{The column name in the data frame for the y-axis variable.}

\item{linewidth}{The width of the line, Default: 0.5}

\item{fill}{The fill color for the points, Default: x-axis variable}

\item{shape}{The shape of the points, Default: 21}

\item{point_size}{The size of the points, Default: 6}

\item{line_color}{The color of the lines, Default: 'black'}

\item{colors}{A vector of colors for fill gradient, Default: NULL (uses pre-defined color palette)}

\item{theme}{The ggplot theme to be used, Default: 'fivethirtyeight'. Other options include 'minimal' and 'classic'}
}
\value{
An object of class 'ggplot' representing the dot plot.
}
\description{
This function generates a dot plot with the given data, x and y. The point's aesthetics and plot theme can be customized.
}
\details{
This function generates a dot plot with the given data, x and y. The points' aesthetics and plot theme can be customized.
}
