% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Create Dotplots of Cognitive Domains}
\usage{
dotplot(data, x, y, sd = NULL, domain = NULL, fill = "domain")
}
\arguments{
\item{data}{Dataset variable.}

\item{x}{X Mean z-score of domain.}

\item{y}{Y Domain/variable name of the y-axis.}

\item{domain}{Optional "domain/area", Default: NULL}

\item{fill}{Optional name of variable to fill points by, Default: 'domain'}

\item{k}{A constant e.g., percentile or z-score to compute uncertainty.}
}
\value{
A dotplot "lollipop" plot of neuropsych domain scores using ggplot2
}
\description{
Make a dotplot for cognitive domains
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{scale_colour_gradient}}, \code{\link[ggplot2]{guide_colourbar}}, \code{\link[ggplot2]{margin}}, \code{\link[ggplot2]{geom_errorbarh}}, \code{\link[ggplot2]{theme}}
\code{\link[stats]{reorder.default}}
\code{\link[ggthemes]{theme_fivethirtyeight}}
}
