% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy_dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Create Dotplots of Cognitive Domains}
\usage{
dotplot(
  data,
  x,
  y,
  domain = NULL,
  fill = NULL,
  size = NULL,
  palette = NULL,
  ...
)
}
\arguments{
\item{data}{Dataset variable.}

\item{x}{X Mean z-score of domain.}

\item{y}{Y Domain/variable name of the y-axis.}

\item{fill}{Optional name of variable to fill points by, Default: 'domain'}

\item{size}{Dot size, Default: NULL}

\item{palette}{Color palette}

\item{...}{other arguments to be passed to
\code{\link[ggplot2]{geom_dotplot}}.}
}
\value{
Produce a dotplot of domain scores
}
\description{
Make a dotplot for cognitive domains.
}
\details{
This is used to collapse individual test scores into composites and
then plot them ...
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{aes}},
\code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{geom_point}},
\code{\link[ggplot2]{scale_colour_gradient}},
\code{\link[ggplot2]{guide_colourbar}}, \code{\link[ggplot2]{margin}},
\code{\link[ggplot2]{theme}}
\code{\link[stats]{reorder.default}}
\code{\link[ggthemes]{theme_fivethirtyeight}}
}
