% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Make a dotplot for cognitive domains}
\usage{
dotplot(data, x, y, ...)
}
\arguments{
\item{data}{dataset}

\item{x}{x-axis}

\item{y}{y-axis}

\item{...}{additional arguments}
}
\value{
A dotplot using ggplot2
}
\description{
Make a dotplot for cognitive domains
}
\examples{
dotplot(iris, iris$Sepal.Length, iris$Species)
}
