% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\title{Dotplot}
\usage{
dotplot(data, x, y, sd = NULL, domain = NULL, fill = "domain")
}
\arguments{
\item{data}{Dataset to import}

\item{x}{X variable}

\item{y}{Y variable}

\item{sd}{PARAM_DESCRIPTION, Default: NULL}

\item{domain}{PARAM_DESCRIPTION, Default: NULL}

\item{fill}{PARAM_DESCRIPTION, Default: 'domain'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{geom_point}}, \code{\link[ggplot2]{scale_colour_gradient}}, \code{\link[ggplot2]{guide_colourbar}}, \code{\link[ggplot2]{margin}}, \code{\link[ggplot2]{geom_errorbarh}}, \code{\link[ggplot2]{theme}}
\code{\link[stats]{reorder.default}}
\code{\link[ggthemes]{theme_fivethirtyeight}}
}
