% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dotplot}
\alias{dotplot}
\title{Create Dotplot for Neurocognitive Domains
This function generates a dotplot for neurocognitive and neurobehavioral
domains.}
\usage{
dotplot(
  data,
  x,
  y,
  linewidth = 0.5,
  fill = x,
  shape = 21,
  point_size = 6,
  line_color = "black",
  colors = NULL,
  theme = "fivethirtyeight",
  return_plot = NULL,
  filename = NULL,
  ...
)
}
\arguments{
\item{data}{The dataset or df containing the data for the dotplot.}

\item{x}{The column name in the data frame for the x-axis variable, typically
the mean z-score for a cognitive domain.}

\item{y}{The column name in the data frame for the y-axis variable, typically
the cognitive domain to plot.}

\item{linewidth}{The width of the line, Default: 0.5}

\item{fill}{The fill color for the points, Default: x-axis variable}

\item{shape}{The shape of the points, Default: 21}

\item{point_size}{The size of the points, Default: 6}

\item{line_color}{The color of the lines, Default: 'black'}

\item{colors}{A vector of colors for fill gradient, Default: NULL (uses
pre-defined color palette)}

\item{theme}{The ggplot theme to be used, Default: 'fivethirtyeight'. Other
options include 'minimal' and 'classic'}

\item{return_plot}{Whether to return the plot object, Default: TRUE}

\item{filename}{The filename to save the plot to, Default: NULL}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
An object of class 'ggplot' representing the dotplot.
}
\description{
Create Dotplot for Neurocognitive Domains
This function generates a dotplot for neurocognitive and neurobehavioral
domains.
}
