% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown.R
\name{drilldown}
\alias{drilldown}
\title{Highcharter Drilldown}
\usage{
drilldown(data, patient, ...)
}
\arguments{
\item{data}{Dataset to use.}

\item{patient}{Patient initials, 4 letters.}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
plot
}
\description{
\code{drilldown} creates a highcharter drilldown interactive plot.
}
