% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown.R
\name{drilldown}
\alias{drilldown}
\title{Drilldown on neuropsych domains.}
\usage{
drilldown(data, patient = NULL, ...)
}
\arguments{
\item{data}{Dataset to use.}

\item{patient}{Patient's first name.}

\item{...}{Numeric, complex, or logical vectors.}
}
\value{
A drilldown plot
}
\description{
This function uses the R Highcharter package and drilldown function to "drill down" on neuropsychological domains and test scores.
}
\details{
\code{drilldown} Creates a highcharter drilldown interactive plot.
}
