% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldown.R
\name{drilldown}
\alias{drilldown}
\title{Drilldown on neuropsych domains.}
\usage{
drilldown(data, mean_z = NULL, mean_pct = NULL)
}
\arguments{
\item{data}{Dataset to use.}

\item{mean_z}{Mean z-score.}

\item{mean_pct}{Mean percentile.}
}
\value{
A drilldown plot
}
\description{
This function uses the R Highcharter package and drilldown function to "drilldown" on neuropsychological domains and test scores.
}
\details{
\code{drilldown} Creates a highcharter drilldown interactive plot.
}
\seealso{
\code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{case_when}}, \code{\link[dplyr]{arrange}}
\code{\link[tibble]{tibble}}
\code{\link[highcharter]{tooltip_table}}, \code{\link[highcharter]{highchart}}, \code{\link[highcharter]{hc_title}}, \code{\link[highcharter]{hc_add_series}}, \code{\link[highcharter]{hcaes}}, \code{\link[highcharter]{hc_xAxis}}, \code{\link[highcharter]{hc_yAxis}}, \code{\link[highcharter]{hc_tooltip}}, \code{\link[highcharter]{hc_plotOptions}}, \code{\link[highcharter]{hc_drilldown}}, \code{\link[highcharter]{hc_add_theme}}, \code{\link[highcharter]{hc_theme_sandsignika}}
}
