% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drilldown}
\alias{drilldown}
\title{Drilldown on neuropsych domains.}
\usage{
drilldown(data, mean_z = NULL, mean_pct = NULL)
}
\arguments{
\item{data}{Dataset to use.}

\item{mean_z}{Mean z-score.}

\item{mean_pct}{Mean percentile.}
}
\value{
A drilldown plot
}
\description{
This function uses the R Highcharter package and drilldown function to "drilldown" on neuropsychological domains and test scores.
}
\details{
\code{drilldown} Creates a highcharter drilldown interactive plot.
}
