% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{extract_and_glue_neuro_text}
\alias{extract_and_glue_neuro_text}
\title{Extracts text from the neuropsych data and writes it to a file.}
\usage{
extract_and_glue_neuro_text(df, filter, result, file = "_raw_text.qmd")
}
\arguments{
\item{df}{The data frame of results}

\item{filter}{The vector of filters to apply}

\item{result}{The resulting string}

\item{file}{The name of the output file, default "_raw_text.qmd"}
}
\value{
The name of the output file
}
\description{
Extracts text from the neuropsych data and writes it to a file.
}
