% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_and_save_transcript.R
\name{extract_and_save_transcript}
\alias{extract_and_save_transcript}
\title{Extract text}
\usage{
extract_and_save_transcript(x, file, begin, end, eol = "\\\\n")
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{file}{PARAM_DESCRIPTION}

\item{begin}{PARAM_DESCRIPTION}

\item{end}{PARAM_DESCRIPTION}

\item{eol}{PARAM_DESCRIPTION, Default: '\\n'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
