% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_and_save_transcript.R
\name{extract_and_save_transcript}
\alias{extract_and_save_transcript}
\title{Extract transcribed text and save as text/md file.}
\usage{
extract_and_save_transcript(x, file, begin, end, eol = "\\\\n")
}
\arguments{
\item{x}{A character vector}

\item{file}{File or connection to write to.}

\item{begin}{First anchor}

\item{end}{Second anchor}

\item{eol}{The line separator. Defaults to ⁠\\.c}
}
\value{
returns x, invisibly.
}
\description{
Extracts text between defined patterns from a specified file and saves the extracted text as a markdown file.
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
