% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_data.R
\name{filter_data}
\alias{filter_data}
\title{Filters Data by Domain and Scale}
\usage{
filter_data(data, domain, scale)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{domain}{The domain name that the user wants to filter by}

\item{scale}{A text file containing a list of scales}
}
\value{
Returns a filtered data frame
}
\description{
This function filters a dataframe by domain and scale.
}
