% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R
\name{gpluck_compute_percentile_range}
\alias{gpluck_compute_percentile_range}
\title{Compute percentile scores and ranges}
\usage{
gpluck_compute_percentile_range(
  .x,
  .score = NA,
  .score_type = c("z_score", "scaled_score", "t_score", "standard_score"),
  percentile = NA,
  range = NA,
  pct1 = NA,
  pct2 = NA,
  pct3 = NA,
  z = NA,
  ...
)
}
\arguments{
\item{.x}{Pronoun for data.frame}

\item{.score}{Known score}

\item{.score_type}{Known type of score}

\item{percentile}{Unknown percentile}

\item{range}{Unknown test score classification range}

\item{pct1}{Temp pct 1}

\item{pct2}{Temp pct 2}

\item{pct3}{Temp pct 3}

\item{z}{Z score}

\item{...}{Additional expressions}
}
\value{
A table with standardized scores
}
\description{
Compute percentile scores and ranges
}
