% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_locate_areas}
\alias{gpluck_locate_areas}
\title{Locate PDF Areas}
\usage{
gpluck_locate_areas(file, pages = NULL, ...)
}
\arguments{
\item{file}{The File name of the input PDF.}

\item{pages}{A vector of character strings specifying which pages to analyse, Default: NULL}

\item{...}{Additional arguments passed to locate_areas() in the tabulapdf package.}
}
\value{
A list containing all areas found by area extraction algorithem.
}
\description{
This is a function to pluck and locate areas from file.
}
\details{
Use this function with caution when handling sensitive PDFs as it involves rasterizing those documents. Also use this function if your PDF contains tables and you want to access those tables programmatically.
}
\seealso{
\code{\link[tabulapdf]{locate_areas}}
}
