% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck.R
\name{gpluck_make_columns}
\alias{gpluck_make_columns}
\title{Make columns for npsych tables}
\usage{
gpluck_make_columns(
  table,
  scale = NULL,
  raw_score = NULL,
  score = NULL,
  range = NULL,
  percentile = NULL,
  ci_95 = NULL,
  test,
  test_name,
  domain = c("Intelligence/General Ability", "Academic Skills", "Verbal/Language",
    "Visual Perception/Construction", "Attention/Executive", "Memory", "Motor",
    "Social Cognition", "Behavioral/Emotional/Social", "Personality Disorders",
    "Psychiatric Disorders", "Substance Use Disorders", "Psychosocial Problems", "ADHD",
    "Executive Dysfunction", "Effort/Validity", NA),
  subdomain = NULL,
  narrow = NULL,
  pass = c("Planning", "Attention", "Sequential", "Simultaneous", "Knowledge", NA),
  verbal = c("Verbal", "Nonverbal", NA),
  timed = c("Timed", "Untimed", NA),
  test_type = c("npsych_test", "rating_scale", "validity_indicator", "item", NA),
  score_type = c("raw_score", "scaled_score", "t_score", "standard_score", "z_score",
    "percentile", "base_rate", "beta_coefficient", NA),
  absort = NULL,
  description = NULL,
  result = NULL,
  ...
)
}
\arguments{
\item{table}{Name of table}

\item{scale}{Name of scale/subtest}

\item{raw_score}{Raw score for scale}

\item{score}{Standardized score}

\item{range}{Test score range}

\item{percentile}{Percentile rank}

\item{ci_95}{95\% CI}

\item{test}{Name of test}

\item{test_name}{Name of test full}

\item{domain}{Cognitive domain}

\item{subdomain}{Cognitive subdomain#'}

\item{narrow}{Narrow cognitive subdomain}

\item{pass}{PASS model area}

\item{verbal}{Verbal or nonverbal test}

\item{timed}{Timed or untimed test}

\item{test_type}{Test type}

\item{score_type}{Score type}

\item{absort}{Sort file}

\item{description}{Description of scale and ability it measures}

\item{result}{Performance on this scale}

\item{...}{Other args}
}
\value{
A table for the report
}
\description{
Make new columns for neuropsych tables more of a description.
}
