% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_make_columns}
\alias{gpluck_make_columns}
\title{Insert Variables and Data from Tables into DF.}
\usage{
gpluck_make_columns(
  data,
  test,
  test_name,
  scale = NULL,
  raw_score = NULL,
  score = NULL,
  ci_95 = NULL,
  percentile = NULL,
  range = NULL,
  domain = c("General Cognitive Ability", "Intelligence/General Ability",
    "Academic Skills", "Verbal/Language", "Visual Perception/Construction",
    "Attention/Executive", "Memory", "Motor", "Social Cognition",
    "Emotional/Behavioral/Personality", "Behavioral/Emotional/Social",
    "Psychiatric Disorders", "Personality Disorders", "Substance Use",
    "Psychosocial Problems", "ADHD", "Executive Functioning", "Adaptive Functioning",
    "Effort/Validity"),
  subdomain = NULL,
  narrow = NULL,
  pass = c("Planning", "Attention", "Sequential", "Simultaneous", "Knowledge", NA),
  verbal = c("Verbal", "Nonverbal", NA),
  timed = c("Timed", "Untimed", NA),
  test_type = c("npsych_test", "rating_scale", "validity_indicator", "item", "basc3", NA),
  score_type = c("raw_score", "scaled_score", "t_score", "standard_score", "z_score",
    "percentile", "base_rate", "beta_coefficient", NA),
  absort = NULL,
  description = NULL,
  result = NULL,
  ...
)
}
\arguments{
\item{data}{Data frame to convert to table and then csv.}

\item{test}{Name of test that information will be extracted from.}

\item{test_name}{Test name as provided in test field.}

\item{scale}{Name of subscale from neuropsych test or battery. Default: NULL}

\item{raw_score}{Raw score if available. Default: NULL}

\item{score}{Standardized Score for the given test. Default: NULL}

\item{ci_95}{Confidence interval level (95\%) for the given test. Default: NULL}

\item{percentile}{Percentile the patient's performance falls at for the given test. Default: NULL}

\item{range}{Range of performance e.g., Below Average, Average, Above
Average, etc the given test or subscale. Default: NULL}

\item{domain}{Domain of the test, e.g. Academic Skills. Default: c("General
Cognitive Ability", "Academic Skills", "Verbal/Language", "Visual
Perception/Construction", "Attention/Executive", "Memory", "Motor", "Social
Cognition", "Emotional/Behavioral/Personality", "Behavioral/Emotional/Social", "Personality Disorders", "Psychiatric Disorders", "Substance Use", "Psychosocial Problems", "ADHD", "Executive Function", "Adaptive Function", "Effort/Validity")}

\item{subdomain}{Cognitive subdomain of the scale. Default: NULL}

\item{narrow}{Narrow cognitive domain of the scale. Default: NULL}

\item{pass}{PASS Cognitive Model. Default: c("Planning", "Attention", "Sequential", "Simultaneous", "Knowledge",
NA)}

\item{verbal}{Type of verbal ability tested, e.g. Verbal, Nonverbal. Default: c("Verbal", "Nonverbal", NA)}

\item{timed}{Indicates if the test is timed or not. Default: c("Timed", "Untimed", NA)}

\item{test_type}{Type of test, e.g. npsych_test, rating_scale,
validity_indicator, item. Default: c("npsych_test", "rating_scale",
"validity_indicator", "basc3", "item", NA)}

\item{score_type}{Type of score reported, e.g. raw_score, scaled_score, t_score, standard_score, z_score, percentile, base_rate, beta_coefficient. Default: c("raw_score", "scaled_score", "t_score", "standard_score", "z_score",
"percentile", "base_rate", "beta_coefficient", NA)}

\item{absort}{Default order of scales to use for sorting. Default: NULL}

\item{description}{Description of the test or task. Default: NULL}

\item{result}{Concatenate the results to include details of test performance. Default: NULL}

\item{...}{Other parameters.}
}
\value{
A modified data frame with additional columns.
}
\description{
This function takes a data frame containing text data from PDF tables, and makes additional columns of binary, range or score values for the specified domain, subdomains, test types, etc.
}
\details{
This function adds new columns to a data frame by extracting numerical values from PDF tables.
}
