% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpluck_make_score_ranges.R
\name{gpluck_make_score_ranges}
\alias{gpluck_make_score_ranges}
\title{gpluck_make_score_ranges}
\usage{
gpluck_make_score_ranges(
  table,
  score = NULL,
  percentile = NULL,
  range,
  test_type,
  ...
)
}
\arguments{
\item{table}{name of table}

\item{score}{standard score.}

\item{percentile}{percentile rank.}

\item{range}{Range of performance score.}

\item{test_type}{Type of test.}

\item{...}{More if needed}
}
\value{
Returns a modfied table.
}
\description{
Make test score ranges
}
\examples{
neurocog <- gpluck_make_score_ranges(
table = neurocog,
 score = 50,
  percentile = 50,
   range = "",
    test_type = "npsych_test")

}
