% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_make_score_ranges}
\alias{gpluck_make_score_ranges}
\title{FUNCTION_TITLE}
\usage{
gpluck_make_score_ranges(
  table,
  score = NULL,
  percentile = NULL,
  range = range,
  test_type = c("npsych_test", "rating_scale", "validity_indicator", "basc3"),
  ...
)
}
\arguments{
\item{table}{PARAM_DESCRIPTION}

\item{score}{PARAM_DESCRIPTION, Default: NULL}

\item{percentile}{PARAM_DESCRIPTION, Default: NULL}

\item{range}{PARAM_DESCRIPTION, Default: range}

\item{test_type}{PARAM_DESCRIPTION, Default: c("npsych_test", "rating_scale", "validity_indicator", "basc3")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
