% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_make_score_ranges}
\alias{gpluck_make_score_ranges}
\title{Make Score Ranges}
\usage{
gpluck_make_score_ranges(
  table,
  score,
  percentile,
  range,
  subdomain = NULL,
  test_type = c("npsych_test", "rating_scale", "performance_validity",
    "symptom_validity", "basc3"),
  ...
)
}
\arguments{
\item{table}{The table containing the lower bound and upper bound score columns}

\item{score}{Optional lower bound or upper bound score column name. If omitted, use \code{Score}}

\item{percentile}{Option percentiles column name. if omitted, use \code{Percentile}}

\item{range}{Score performance range. if omitted, use \code{Range}}

\item{subdomain}{Which subdomain.}

\item{test_type}{A vector of test types to consider. Default: c("npsych_test", "rating_scale", "validity_indicator", "basc3")}

\item{...}{Other arguments passed on to \code{dplyr::filter}}
}
\value{
An object that matches the input but with the ability to get score ranges for any given test type
}
\description{
This function takes a text table of score ranges for each percentile, and returns an object that matches the input but with the ability to get score ranges for any given test type
}
\details{
This function takes a text table of score ranges for each percentile, and returns an object that matches the input but with the ability to get score ranges for any given test type
}
