% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_dotplot_typ}
\alias{make_dotplot_typ}
\title{Make a dotplot for cognitive domain scores
This function creates a dotplot figure for a given dataset.}
\usage{
make_dotplot_typ(
  data,
  pheno,
  x = data$z_mean_subdomain,
  y = data$subdomain,
  colors = NULL,
  return_plot = TRUE,
  filename = c("pdf", "png", "svg")
)
}
\arguments{
\item{data}{Dataframe containing the input data}

\item{pheno}{Phenotype of interest}

\item{x}{Variable plotted on the x-axis, by default data_dotplot$z_mean_narrow}

\item{y}{Variable plotted on the y-axis, by default data_dotplot$narrow}

\item{colors}{A vector of colors to be used for the plot}
}
\value{
an object of class figure object
}
\description{
Make a dotplot for cognitive domain scores
This function creates a dotplot figure for a given dataset.
}
