% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_fig}
\alias{make_fig}
\title{Make a figure}
\usage{
make_fig(data, x = data_dotplot$z_mean_narrow, y = data_dotplot$narrow, pheno)
}
\arguments{
\item{data}{Dataframe containing the input data}

\item{x}{Variable plotted on the x-axis, by default data_dotplot$z_mean_narrow}

\item{y}{Variable plotted on the y-axis, by default data_dotplot$narrow}

\item{pheno}{Phenotype of interest}
}
\value{
an object of class figure object
}
\description{
Make a figure
}
