% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{make_tbl_md_typ}
\alias{make_tbl_md_typ}
\title{Create a Table for Typst
This function creates a table of descriptive statistics for a given dataset.}
\usage{
make_tbl_md_typ(data)
}
\arguments{
\item{data}{A dataframe containing the columns to be used in the table.}
}
\value{
tbl_md A table of descriptive statistics.
}
\description{
Create a Table for Typst
This function creates a table of descriptive statistics for a given dataset.
}
