% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_tibble}
\alias{make_tibble}
\title{Make a tibble for plots}
\usage{
make_tibble(
  data,
  pheno = NULL,
  domain = NULL,
  columns = c("scale", "score", "percentile", "range", "subdomain", "test_name"),
  percentile = NULL,
  names = c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test"),
  tibb = NULL,
  ...
)
}
\arguments{
\item{data}{A character string indicating the name of the dataset, Default: '.'}

\item{pheno}{A character vector indicating the domain of interest, Default: NULL}

\item{domain}{A character string indicating the domain of interest, Default: NULL}

\item{columns}{A character vector of column names to be included in the tibble, Default: c("scale", "score", "percentile", "range", "subdomain", "test_name")}

\item{percentile}{A numeric vector indicating the percentile values to be truncated, Default: NULL}

\item{names}{A character vector of column names to be set in the tibble, Default: c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test")}

\item{tibb}{A data frame containing the data, Default: NULL}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A tibble with columns specified in 'columns', and set column names to 'names'.
}
\description{
Create a tibble with columns specified in 'columns', and set column names to 'names'.
}
\details{
The function filters the data by 'domain', selects columns from 'columns', and sets the column names to 'names'. It then truncates the values of 'percentile' column.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}

}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}
\link[tidyselect:all_of]{all_of}
\link[purrr:reexports]{set_names}
}
