% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tibble.R
\name{make_tibble}
\alias{make_tibble}
\title{FUNCTION_TITLE}
\usage{
make_tibble(
  tibb,
  data = ".",
  pheno = NULL,
  domain = NULL,
  columns = c("scale", "score", "percentile", "range", "subdomain", "test_name"),
  percentile = NULL,
  names = c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test"),
  ...
)
}
\arguments{
\item{tibb}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION, Default: '.'}

\item{pheno}{PARAM_DESCRIPTION, Default: NULL}

\item{domain}{PARAM_DESCRIPTION, Default: NULL}

\item{columns}{PARAM_DESCRIPTION, Default: c("scale", "score", "percentile", "range", "subdomain", "test_name")}

\item{percentile}{PARAM_DESCRIPTION, Default: NULL}

\item{names}{PARAM_DESCRIPTION, Default: c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:filter]{filter}, \link[dplyr:select]{select}, \link[dplyr:mutate]{mutate}
\link[tidyselect:all_of]{all_of}
\link[purrr:reexports]{set_names}
}
