% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tibble.R
\name{make_tibble}
\alias{make_tibble}
\title{Make tibble for npsych report tables}
\usage{
make_tibble(
  tibb,
  data = c("neurocog", "neurobehav"),
  pheno = NULL,
  domain = NULL,
  columns = c("scale", "score", "percentile", "range", "subdomain", "test_name"),
  percentile = NULL,
  digits = 0,
  names = c("Scale", "Score", "&#x2030 Rank", "Range", "Subdomain", "Test"),
  ...
)
}
\arguments{
\item{tibb}{name of tibble to make}

\item{data}{dataset}

\item{pheno}{which phenotype/cog trait}

\item{domain}{which cognitive domain}

\item{columns}{columns to keep}

\item{percentile}{percentile score}

\item{digits}{how many digits to round}

\item{names}{names of variables/columns}

\item{...}{additional arguments}
}
\value{
A dotplot using ggplot2
}
\description{
Make tibble for npsych report tables
}
\examples{
require(magrittr)
make_tibble(tibb = iq, data = neurocog, pheno = "Intelligence/General Ability")
}
