% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tibble.R
\name{make_tibble}
\alias{make_tibble}
\title{Make tibble table for npsych report tables}
\usage{
make_tibble(
  tibb,
  data = ".",
  pheno = NULL,
  domain = NULL,
  columns = c("scale", "score", "percentile", "range", "subdomain", "test_name"),
  percentile = NULL,
  names = c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test"),
  ...
)
}
\arguments{
\item{tibb}{Name of tibble to make}

\item{data}{Dataset to use}

\item{pheno}{Which phenotype/cog trait}

\item{domain}{Which cognitive domain}

\item{columns}{Columns to keep}

\item{percentile}{Percentile score}

\item{names}{Names of variables/columns}

\item{...}{Additional arguments}
}
\value{
A tibble for plotting
}
\description{
Make a tibble table for reports.
}
\examples{
make_tibble(
  tibb = iq,
  data = neurocog,
  pheno = "Intelligence/General Ability"
)

}
