% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_tibble}
\alias{make_tibble}
\title{Make a Tibble for Plots}
\usage{
make_tibble(
  data,
  domain,
  columns = c("scale", "score", "percentile", "range", "subdomain", "test_name"),
  percentile = NULL,
  round = 0,
  names = c("Scale", "Score", "‰ Rank", "Range", "Subdomain", "Test"),
  ...
)
}
\arguments{
\item{data}{A data.frame/dataset, Default: '.'}

\item{domain}{A character string indicating the domain of interest, Default: NULL}

\item{columns}{A character vector of column names to be included in the tibble, Default: c("scale", "score", "percentile", "range", "subdomain", "test_name")}

\item{percentile}{A numeric vector indicating the percentile values to be truncated, Default: NULL}

\item{round}{A numeric vector indicating the number of digits to round to, Default: 0}

\item{names}{A character vector of column names to be set in the tibble, Default: NULL}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A tibble with columns specified in 'columns', and set column names to 'names'.
}
\description{
Create a tibble with columns specified in 'columns', and set column names to 'names'.
}
