% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{pct}
\alias{pct}
\title{Convert T scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks. That is about it.}
\usage{
pct(x, m, sd)
}
\arguments{
\item{x}{Raw test score}

\item{m}{Mean of whatever type of score}

\item{sd}{Standard deviation (SD) of whatever type of score}
}
\value{
A percentile rank score
}
\description{
Convert T scores to z scores to percentiles.
This is a functioning that converts T-scores to z-scores to percentile
ranks. That is about it.
}
\examples{
pct(x = 103, m = 100, sd = 15)

}
