% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{pred}
\alias{pred}
\title{Predicted test score}
\usage{
pred(x, y, r, m, ...)
}
\arguments{
\item{x}{Score 1 (ability)}

\item{y}{Score 2 (achievement)}

\item{r}{Correlation between ability test and achievement test}

\item{m}{Mean of test score}

\item{...}{Additional arguments}
}
\value{
A predicted score.
}
\description{
Estimate the difference between an ability and achievement score.
}
\examples{
pred(x = 122, y = 84, r = .65, m = 100)

}
