% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_neuropsych.R
\name{read_data}
\alias{read_data}
\title{Read Neuropsych Data for Subsetting}
\usage{
read_data(pheno, domains, scales, ...)
}
\arguments{
\item{pheno}{Character vector for phenotype. Options are "adhd" or "emotion".}

\item{domains}{Character vector for domain. Options are "neurocog" or "neurobehav".}

\item{scales}{Character vector for scale.}

\item{...}{Additional options.}
}
\value{
A tibble containing various subsets of the data.
}
\description{
Read Neuropsych Data for Subsetting
}
