% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_neuropsych.R
\name{read_data}
\alias{read_data}
\title{Read Neuropsych Data for Subsetting}
\usage{
read_data(pheno)
}
\arguments{
\item{pheno}{Character vector for phenotype. Options are "adhd" or "emotion".}
}
\value{
A tibble containing various subsets of the data.
}
\description{
Read Neuropsych Data for Subsetting
}
