% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typst.R
\name{read_data}
\alias{read_data}
\title{Read data from a csv file.}
\usage{
read_data(pheno)
}
\arguments{
\item{pheno}{Character vector for phenotype. Options are "adhd" or "emotion".}
}
\value{
A tibble containing the data.
}
\description{
Read data from a csv file.
}
