% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_scale.R
\name{rename_scale}
\alias{rename_scale}
\title{FUNCTION_TITLE}
\usage{
rename_scale(table, scale, name1, name2)
}
\arguments{
\item{table}{PARAM_DESCRIPTION}

\item{scale}{PARAM_DESCRIPTION}

\item{name1}{PARAM_DESCRIPTION}

\item{name2}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\link[dplyr:mutate]{mutate}
\link[stringr:str_replace]{str_replace}
}
