% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_scale.R
\name{rename_scale}
\alias{rename_scale}
\title{Rename neuropsych table scale variable}
\usage{
rename_scale(table, scale, name1, name2)
}
\arguments{
\item{table}{Name of table}

\item{scale}{Name of scale}

\item{name1}{Old name}

\item{name2}{New name}
}
\description{
Rename neuropsych table scale variable
}
