% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{FUNCTION_TITLE}
\usage{
render_report(
  patient = "Biggie",
  input,
  output_format = NULL,
  params = NULL,
  ...
)
}
\arguments{
\item{patient}{PARAM_DESCRIPTION}

\item{input}{PARAM_DESCRIPTION}

\item{output_format}{adult or child}

\item{params}{Parameters or "ask"}

\item{...}{Patient name, etc}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
}
}
}
\seealso{
\link[rmarkdown:render]{render}
\link[bookdown:html_document2]{pdf_document2}
}
