% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocft.R
\name{rocft_copy}
\alias{rocft_copy}
\title{ROCFT Copy Standardized Score}
\usage{
rocft_copy(raw_score, age)
}
\arguments{
\item{raw_score}{Patients raw score}

\item{age}{Patients age}
}
\value{
A printed T-Score of ROCFT Copy performance
}
\description{
ROCFT predicted scores, for age range of 22 -79 years, for copy trial.
}
\examples{
rocft_copy_score <- rocft_copy(raw_score = 32, age = 24)
}
