% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_scoring.R
\name{rocft_copy}
\alias{rocft_copy}
\title{Rey Complex Figure, Copy}
\usage{
rocft_copy(raw_score, age)
}
\arguments{
\item{raw_score}{Description for the raw_score parameter.}

\item{age}{Description for the age parameter.}

\item{t_score}{Description for the t_score parameter.}
}
\value{
Description for the return value of the function.
}
\description{
Provide a brief description of the rocft_copy function.
}
\details{
Provide additional details about the function if needed.
}
