% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropsych_test_scoring.R
\name{rocft_delay}
\alias{rocft_delay}
\title{Rey Complex Figure, Delayed Recall}
\usage{
rocft_delay(raw_score, age)
}
\arguments{
\item{raw_score}{Description for the raw_score parameter.}

\item{age}{Description for the age parameter.}
}
\value{
Description for the return value of the function.
}
\description{
Provide a brief description of the rocft_delay function.
}
