% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{save_as_markdown}
\alias{save_as_markdown}
\title{Save Speech-to-Text As Markdown}
\usage{
save_as_markdown(text, output_path)
}
\arguments{
\item{text}{Name of file.}

\item{output_path}{Where to save outoput.}
}
\value{
Markdown text file
}
\description{
This function saves the extracted text from a speech-to-text as markdown.
}
