% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_gt}
\alias{tbl_gt}
\title{Make GT table per domain}
\usage{
tbl_gt(
  data,
  table_name = NULL,
  source_note = NULL,
  title = NULL,
  tab_stubhead = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{File or path to data.}

\item{table_name}{Name of the table to be saved.}

\item{source_note}{Source note to be added to the table.}

\item{title}{Title of the table.}

\item{tab_stubhead}{Stubhead of the table.}

\item{caption}{Caption of the table.}
}
\value{
A formatted table with domain counts.
}
\description{
Create a table of domain counts using dplyr and gt packages.
}
\details{
This function creates a table of domain counts from a data frame using the dplyr and gt packages. It also saves the table with the specified name.
}
