% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-tables.R
\name{tbl_gt}
\alias{tbl_gt}
\title{tbl_gt}
\usage{
tbl_gt(file_path, filter, domain, table_name)
}
\arguments{
\item{file_path}{Path to the data frame.}

\item{filter}{Filter for the data frame.}

\item{domain}{Domain to be used in the data frame.}

\item{table_name}{Name of the table.}
}
\value{
A formatted table with domain counts.
}
\description{
Create a table of domain counts using dplyr and gt packages.
}
\details{
This function creates a table of domain counts from a data frame using the dplyr and gt packages. It also saves the table with the specified name.
}
\examples{
\dontrun{
if (interactive()) {
  # EXAMPLE1
  df_domain <- read.csv("data/domain_counts.csv")
  tbl_gt(df_domain, filter = TRUE, domain = "test", table_name = "table_domain")
}
}
}
\seealso{
\code{\link[dplyr]{mutate}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{select}}
\code{\link[gt]{gt}}, \code{\link[gt]{cols_label}}, \code{\link[gt]{tab_stub_indent}}, \code{\link[gt]{tab_header}}, \code{\link[gt]{sub_missing}}, \code{\link[gt]{tab_options}}, \code{\link[gt]{cols_align}}, \code{\link[gt]{tab_source_note}}, \code{\link[gt]{character(0)}}, \code{\link[gt]{gtsave}}
}
