% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_gt}
\alias{tbl_gt}
\title{Make Table Using \code{gt} Package for Neurocognitive Domains}
\usage{
tbl_gt(
  data,
  pheno = NULL,
  table_name = NULL,
  source_note = NULL,
  names = NULL,
  title = NULL,
  tab_stubhead = NULL,
  caption = NULL,
  process_md = FALSE,
  fn_scaled_score = NULL,
  fn_standard_score = NULL,
  fn_t_score = NULL,
  fn_z_score = NULL,
  fn_raw_score = NULL,
  grp_scaled_score = NULL,
  grp_standard_score = NULL,
  grp_t_score = NULL,
  grp_z_score = NULL,
  grp_raw_score = NULL,
  dynamic_grp = NULL,
  vertical_padding = NULL,
  multiline = TRUE,
  ...
)
}
\arguments{
\item{data}{File or path to data.}

\item{pheno}{Phenotype name.}

\item{table_name}{Name of the table to be saved.}

\item{source_note}{Source note to be added to the table.}

\item{names}{Names of the columns.}

\item{title}{Title of the table.}

\item{tab_stubhead}{Stubhead of the table.}

\item{caption}{Caption of the table.}

\item{process_md}{Process markdown.}

\item{fn_scaled_score}{Footnote for scaled score.}

\item{fn_standard_score}{Footnote for standard score.}

\item{fn_t_score}{Footnote for t score.}

\item{fn_z_score}{Footnote for z score.}

\item{fn_raw_score}{Footnote for raw scores.}

\item{grp_scaled_score}{Groups for scaled score.}

\item{grp_standard_score}{Groups for standard score.}

\item{grp_t_score}{Groups for t score.}

\item{grp_z_score}{Groups for z score.}

\item{grp_raw_score}{Groups for raw scores.}

\item{dynamic_grp}{Generalized grouping parameter.}

\item{vertical_padding}{Vertical padding.}

\item{multiline}{Multiline footnotes, Default = TRUE.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A formatted table with domain counts.
}
\description{
Create a table of domain counts using dplyr and gt packages.
}
