% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_md_typ}
\alias{tbl_md_typ}
\title{Create Markdown Table for Typst}
\usage{
tbl_md_typ(data, caption = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{caption}{A caption for the table.}
}
\value{
A markdown table.
}
\description{
Create a markdown table from a data frame for Typst.
}
\details{
This function creates a markdown table from a data frame.
}
