% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt.R
\name{tmtA}
\alias{tmtA}
\title{Trails A Standardized Score}
\usage{
tmtA(rawScore, age)
}
\arguments{
\item{rawScore}{Patients raw score}

\item{age}{Patients age}
}
\value{
A printed score
}
\description{
Trails A Standardized Score
}
\examples{
tmt_a_tscore <- tmtA(rawScore = 25, age = 21)
}
