% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt.R
\name{tmtB}
\alias{tmtB}
\title{Trails B Standardized Score}
\usage{
tmtB(raw_score, age)
}
\arguments{
\item{raw_score}{Patients raw score (seconds) on Trails B}

\item{age}{Patient's age}
}
\value{
A tscore of TMTB
}
\description{
Trails B Standardized Score
}
\examples{
tmt_b_tscore <- tmtB(raw_score = 75, age = 21)
}
